requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
hdbscan>=0.8.38
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.13.0
groq>=0.9.0
pinecone-client>=3.2.0
lxml>=5.2.0
beautifulsoup4>=4.12.0
requests>=2.32.0
diskcache>=5.6.0
tenacity>=8.3.0
aiohttp>=3.9.0
tqdm>=4.66.0
streamlit>=1.35.0
plotly>=5.22.0
